[#12]fix : 마이페이지 shadow 수정
