+++ src/types/next-auth.d.ts
@@ -35,6 +35,8 @@ declare module 'next-auth/jwt' {
     trackId?: number;
     trackRole?: trackRole;
     trackName?: TrackType;
+    phone?: string;
+    email?: string;
     loginPeriod?: number;
     roleApply?: boolean;
     nickname?: string;
