+++ tailwind.config.ts
@@ -160,6 +160,7 @@ const config = {
       archiveShadow: '0px 4px 100px 0px rgba(185, 210, 225, 0.80)',
       hoverPostCard: '0px 4px 24px 0px rgba(0, 0, 0, 0.15)',
       myPageShadow: '0px 4px 10px 0px rgba(12, 12, 13, 0.10)',
+      myPageBottomShadow: '0px 6px 6px 0px rgba(12, 12, 13, 0.10)',
     },
     backgroundImage: {
       'custom-gradient':
